Bump dependencies, fix clippy issues

+++ programs/twamm/src/oracle.rs
@@ -116,7 +116,7 @@ pub fn get_test_price(
 
     Ok(OraclePrice {
         // price is i64 and > 0 per check above
-        price: oracle_acc.price as u64,
+        price: oracle_acc.price,
         exponent: oracle_acc.expo,
     })
 }
